Move Slack adapter code to pkg/adapter and cmd/
