+++ cmd/slacksource-adapter/Dockerfile
@@ -1,0 +1,36 @@
+FROM golang:1.14-stretch AS builder
+
+ENV CGO_ENABLED 0
+ENV GOOS linux
+ENV GOARCH amd64
+
+WORKDIR /go/src/slacksource-adapter
+
+COPY go.mod go.sum ./
+RUN go mod download
+
+COPY . .
+RUN BIN_OUTPUT_DIR=/bin make slacksource-adapter && \
+    mkdir /kodata && \
+    mv .git/* /kodata/ && \
+    rm -rf ${GOPATH} && \
+    rm -rf ${HOME}/.cache
+
+FROM registry.access.redhat.com/ubi8/ubi-minimal
+
+LABEL name "Triggermesh event source adapter for Slack"
+LABEL vendor "Triggermesh"
+LABEL version "v0.1.0"
+LABEL release "1"
+LABEL summary "The Triggermesh event source adapter for Slack"
+LABEL description "This is the Triggermesh Slack event source adapter for Knative"
+
+# Emulate ko builds
+# https://github.com/google/ko/blob/v0.5.0/README.md#including-static-assets
+ENV KO_DATA_PATH /kodata
+
+COPY --from=builder /kodata/ ${KO_DATA_PATH}/
+COPY --from=builder /bin/slacksource-adapter /
+COPY licenses/ /licenses/
+
+ENTRYPOINT ["/slacksource-adapter"]
