+++ pkg/adapter/slacksource/event.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package adapter
+package slacksource
 
 // SlackEvent contains the event payload
 type SlackEvent map[string]interface{}
