+++ pkg/adapter/slacksource/signing.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package adapter
+package slacksource
 
 import (
 	"crypto/hmac"
