+++ pkg/adapter/slacksource/slack_test.go
@@ -18,7 +18,7 @@ limitations under the License.
 // 	cloudevents "github.com/cloudevents/sdk-go/v2"
 // )
 
-package adapter
+package slacksource
 
 import (
 	"io"
