--- slack/cmd/adapter/Dockerfile
@@ -1,36 +1,0 @@
-FROM golang:1.14-stretch AS builder
-
-ENV CGO_ENABLED 0
-ENV GOOS linux
-ENV GOARCH amd64
-
-RUN apt update && \
-    apt install -y ca-certificates && \
-    rm -rf /var/lib/apt/lists/*
-
-WORKDIR /go/src/sources
-
-COPY go.mod go.sum ./
-RUN go mod download
-
-# NOTE: try not to modify the instructions above to keep cached layers
-
-COPY ./ .
-RUN BIN_OUTPUT_DIR=/bin make -C slack adapter && \
-    rm -rf ${GOPATH} && \
-    rm -rf ${HOME}/.cache
-
-FROM registry.access.redhat.com/ubi8/ubi-minimal
-
-LABEL name "Triggermesh Event Source Adapter for Slack"
-LABEL vendor "Triggermesh"
-LABEL version "v0.1.0"
-LABEL release "1"
-LABEL summary "The Triggermesh Event Source Adapter for Slack"
-LABEL description "This is the Triggermesh Knative Event Source Adapter for Slack"
-
-COPY licenses/ /licenses/
-COPY --from=builder /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
-COPY --from=builder /bin/slack-adapter /slack-adapter
-
-ENTRYPOINT ["/slack-adapter"]
