--- slack/cmd/adapter/kodata/refs
@@ -1,1 +1,0 @@
-../../../../.git/refs
